// service/types.gen.ts
export type ShopResponseSchema = {
    shopId: string;
    shop_no?: string;
    trademark: string;
    branch: string | null;
    total_space: number;
    put_space: number;
    price_base: number;
    verified: boolean;
    displayed: boolean;
    type: string;
    type_tag: string | null;
    photo: Array<string>;
    remark: string | null;
    business_hours: Array<number>;
    total_area: number | null;
    customer_area: number | null;
    clerk_count: number | null;
    business_type: string;
    duration: string;
    sex: string;
    age: Array<number>;
    id_tag: string | null;
    sign_photo: string | null;
    contact_type: string;
    /**
     * 商家简介
     */
    shop_description?: string;
    /**
     * 投放简介
     */
    put_description?: string;
    /**
     * 分类标签
     */
    classify_tag?: string;
    positionId?: string;
    position_no?: string;
    /**
     * 休息日，1-周一 2-周二 3-周三 4-周四 5-周五 6-周六 7-周日 8-按需
     */
    rest_days: Array<string>;
    /**
     * 客流高峰，1-早餐 2-午餐 3-晚餐 4-宵夜 5-上午 6-下午 7-晚上 8-深夜
     */
    volume_peak: Array<string>;
    /**
     * 1-春 2-夏 3-秋 4-冬 5-节假日 6-工作日 7-非工作日
     */
    season: Array<string>;
};
// ...
export type PostShopAddData = {
    body?: {
        /**
         * 商圈ID
         */
        cbdId: string;
        /**
         * 分区ID
         */
        partId: string;
        /**
         * 铺位ID
         */
        positionId?: string;
        /**
         * 类型，1-餐饮 2-轻食 3-茶楼 4-茶饮/咖啡 5-咖啡馆 6-酒店
         */
        type: string;
        /**
         * 品类标签
         */
        type_tag?: string;
        /**
         * 商业类型，1-独立自营店 2-连锁自营店 3-连锁加盟店
         */
        business_type: string;
        /**
         * 字号
         */
        trademark: string;
        /**
         * 分店
         */
        branch?: string;
        /**
         * 坐标，经纬度
         */
        location: [
            number,
            number
        ];
        /**
         * 是否认证
         */
        verified?: boolean;
        /**
         * 经营时长，1-一年内新店 2-1~2年 3-2~5年 4-五年以上
         */
        duration: string;
        /**
         * 是否展示消费数据
         */
        consume_display?: boolean;
        /**
         * 人均消费最低到最高值，单位(分)
         */
        average_expense: [
            number,
            number
        ];
        /**
         * 性别，1-不限 2-男 3-女
         */
        sex: string;
        /**
         * 年龄段最低到最高值
         */
        age: [
            number,
            number
        ];
        /**
         * 身份标签
         */
        id_tag?: string;
        /**
         * 标识图片
         */
        sign_photo?: string;
        /**
         * 认证图片
         */
        verify_photo?: Array<string>;
        /**
         * 外景图片
         */
        environment_photo?: Array<string>;
        /**
         * 内景图片
         */
        building_photo?: Array<string>;
        /**
         * 品牌营销图片
         */
        brand_photo?: Array<string>;
        /**
         * 联系人姓名
         */
        contact_name?: string;
        /**
         * 联系人电话
         */
        contact_phone?: string;
        /**
         * 联系人类型，1-老板 2-店长 3-店员 4-总店管理人员
         */
        contact_type?: string;
        /**
         * 面积，单位(平方米)
         */
        total_area?: number;
        /**
         * 客区面积
         */
        customer_area?: number;
        /**
         * 店员人数
         */
        clerk_count?: number;
        /**
         * 营业时间
         */
        business_hours: [
            number,
            number
        ];
        /**
         * 休息日，1-周一 2-周二 3-周三 4-周四 5-周五 6-周六 7-周日 8-按需
         */
        rest_days: Array<string>;
        /**
         * 客流高峰，1-早餐 2-午餐 3-晚餐 4-宵夜 5-上午 6-下午 7-晚上 8-深夜
         */
        volume_peak: Array<string>;
        /**
         * 1-春 2-夏 3-秋 4-冬 5-节假日 6-工作日 7-非工作日
         */
        season: Array<string>;
        /**
         * 商家简介
         */
        shop_description?: string;
        /**
         * 投放简介
         */
        put_description?: string;
        /**
         * 是否开放
         */
        displayed?: boolean;
        /**
         * 价格基数（单位：分）
         */
        price_base: number;
        /**
         * 分类标签
         */
        classify_tag?: string;
        /**
         * 备注
         */
        remark?: string;
    };
    path?: never;
    query?: never;
    url: '/shop/add';
};

export type PostShopAddResponses = {
    /**
     * 成功响应
     */
    200: {
        code?: number;
        data?: {
            /**
             * 新建商家ID
             */
            id: string;
        };
    };
};

export type PostShopAddResponse = PostShopAddResponses[keyof PostShopAddResponses];

export type PostShopUpdateData = {
    body?: {
        /**
         * 类型，1-餐饮 2-轻食 3-茶楼 4-茶饮/咖啡 5-咖啡馆 6-酒店
         */
        type: string;
        /**
         * 品类标签
         */
        type_tag?: string;
        /**
         * 商业类型，1-独立自营店 2-连锁自营店 3-连锁加盟店
         */
        business_type: string;
        /**
         * 字号
         */
        trademark: string;
        /**
         * 分店
         */
        branch?: string;
        /**
         * 坐标，经纬度
         */
        location: [
            number,
            number
        ];
        /**
         * 是否认证
         */
        verified?: boolean;
        /**
         * 经营时长，1-一年内新店 2-1~2年 3-2~5年 4-五年以上
         */
        duration: string;
        /**
         * 是否展示消费数据
         */
        consume_display?: boolean;
        /**
         * 人均消费最低到最高值，单位(分)
         */
        average_expense: [
            number,
            number
        ];
        /**
         * 性别，1-不限 2-男 3-女
         */
        sex: string;
        /**
         * 年龄段最低到最高值
         */
        age: [
            number,
            number
        ];
        /**
         * 身份标签
         */
        id_tag?: string;
        /**
         * 标识图片
         */
        sign_photo?: string;
        /**
         * 认证图片
         */
        verify_photo?: Array<string>;
        /**
         * 外景图片
         */
        environment_photo?: Array<string>;
        /**
         * 内景图片
         */
        building_photo?: Array<string>;
        /**
         * 品牌营销图片
         */
        brand_photo?: Array<string>;
        /**
         * 联系人姓名
         */
        contact_name?: string;
        /**
         * 联系人电话
         */
        contact_phone?: string;
        /**
         * 联系人类型，1-老板 2-店长 3-店员 4-总店管理人员
         */
        contact_type?: string;
        /**
         * 面积，单位(平方米)
         */
        total_area?: number;
        /**
         * 客区面积
         */
        customer_area?: number;
        /**
         * 店员人数
         */
        clerk_count?: number;
        /**
         * 营业时间
         */
        business_hours: [
            number,
            number
        ];
        /**
         * 休息日，1-周一 2-周二 3-周三 4-周四 5-周五 6-周六 7-周日 8-按需
         */
        rest_days: Array<string>;
        /**
         * 客流高峰，1-早餐 2-午餐 3-晚餐 4-宵夜 5-上午 6-下午 7-晚上 8-深夜
         */
        volume_peak: Array<string>;
        /**
         * 1-春 2-夏 3-秋 4-冬 5-节假日 6-工作日 7-非工作日
         */
        season: Array<string>;
        /**
         * 商家简介
         */
        shop_description?: string;
        /**
         * 投放简介
         */
        put_description?: string;
        /**
         * 是否开放
         */
        displayed?: boolean;
        /**
         * 价格基数（单位：分）
         */
        price_base: number;
        /**
         * 分类标签
         */
        classify_tag?: string;
        /**
         * 备注
         */
        remark?: string;
        /**
         * 商家ID
         */
        id: string;
    };
    path?: never;
    query?: never;
    url: '/shop/update';
};